@@ -11,25 +11,26 @@
     "@testing-library/user-event": "^7.1.2",
     "@types/jest": "^27.4.0",
     "@types/node": "^17.0.12",
-    "@types/react-dom": "^17.0.11",
+    "@types/react-dom": "^18.0.5",
     "@types/react-redux": "^7.1.22",
     "axios": "^0.20.0",
     "lodash": "^4.17.21",
     "node-sass": "^7.0.1",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-hook-form": "^7.24.0",
     "react-icons": "^4.3.1",
     "react-query": "^3.34.8",
     "react-redux": "^7.2.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "^5.0.0",
+    "react-select": "^5.3.2",
     "react-toastify": "^8.1.0",
     "redux": "^4.0.5",
     "redux-devtools-extension": "^2.13.8",
     "redux-thunk": "^2.3.0",
     "sass": "^1.45.2",
-    "styled-components": "^5.3.3",
+    "styled-components": "^5.3.5",
     "typescript": "^4.5.5"
   },
   "author": "ABee Tech",
