@@ -1,10 +1,13 @@
-import React from 'react';
-import { createTodoCategory } from '../../redux/actions/todoCategory.actions';
-import { useDispatch } from 'react-redux';
-import { PrimaryButton } from '../../styles/styles';
-import { useForm } from 'react-hook-form';
-import FormInput from '../FormInput/FormInput';
-import ColorSelect from '../ColorSelect/ColorSelect';
+import React from "react";
+import { createTodoCategory } from "../../redux/actions/todoCategory.actions";
+import { useDispatch } from "react-redux";
+import { PrimaryButton } from "../../styles/styles";
+import { useForm } from "react-hook-form";
+import FormInput from "../FormInput/FormInput";
+import ColorSelect, {
+  ColorSelectWithControl,
+  IColorSelectOptions,
+} from "../ColorSelect/ColorSelect";
 
 interface IAddTodoCategoryProps {
   setOpen: React.Dispatch<React.SetStateAction<boolean>>;
@@ -12,7 +15,7 @@ interface IAddTodoCategoryProps {
 
 interface IFormData {
   name?: string;
-  color?: string;
+  color?: IColorSelectOptions;
 }
 
 const AddTodoCategory = ({ setOpen }: IAddTodoCategoryProps) => {
@@ -23,15 +26,16 @@ const AddTodoCategory = ({ setOpen }: IAddTodoCategoryProps) => {
     register,
     handleSubmit,
     formState: { errors },
+    control,
   } = useForm();
 
   //submit form
   const formSubmitHandler = ({ name, color }: IFormData) => {
     const data = {
       name,
-      color,
+      color: color?.value,
     };
-    console.log(data, 'data');
+    console.log(data, "data");
     dispatch(createTodoCategory(data));
     setOpen(false);
   };
@@ -44,27 +48,27 @@ const AddTodoCategory = ({ setOpen }: IAddTodoCategoryProps) => {
           id="name"
           placeholder="Todo Category"
           errorText={errors?.name?.message}
-          {...register('name', {
-            required: 'Todo category is required',
+          {...register("name", {
+            required: "Todo category is required",
           })}
         />
       </fieldset>
       <fieldset>
-        <ColorSelect
+        <ColorSelectWithControl
+          control={control}
+          id="color"
           options={[
-            { label: 'Vivid Malachite', value: '#07bc0c' },
-            { label: 'Carmine Pink', value: '#e74c3c' },
-            { label: 'Tufts Blue', value: '#3498db' },
-            { label: 'Royal Pink', value: '#e84393' },
-            { label: 'Jonquil', value: '#f1c40f' },
-            { label: 'UFO Green', value: '#2ecc71' },
-            { label: 'Deep Lilac', value: '#9b59b6' },
+            { label: "Vivid Malachite", value: "#07bc0c" },
+            { label: "Carmine Pink", value: "#e74c3c" },
+            { label: "Tufts Blue", value: "#3498db" },
+            { label: "Royal Pink", value: "#e84393" },
+            { label: "Jonquil", value: "#f1c40f" },
+            { label: "UFO Green", value: "#2ecc71" },
+            { label: "Deep Lilac", value: "#9b59b6" },
           ]}
           label="Color"
-          defaultValue="#07bc0c"
-          {...register('color', {
-            required: 'Color is required',
-          })}
+          rules={{ required: "Color is required" }}
+          errorText={errors?.color?.message}
         />
       </fieldset>
       <PrimaryButton
