@@ -1,7 +1,9 @@
+import SelectWithLabel from "../../components/SelectWithLabel/SelectWithLabel";
 import React, { useEffect, useState } from "react";
-import { ErrorMessage } from "src/styles/styles";
+import { Controller } from "react-hook-form";
+import { MdOutlineInvertColorsOff } from "react-icons/md";
 
-interface IColorSelectOptions {
+export interface IColorSelectOptions {
   label: string;
   value: string;
 }
@@ -9,79 +11,72 @@ interface IColorSelectOptions {
 interface IColorSelectProps {
   label?: string;
   options?: IColorSelectOptions[];
-  defaultValue?: string;
   errorText?: string;
+  value?: IColorSelectOptions;
   onChange?: (e: any) => void;
-  [x: string]: any;
 }
 
-const ColorSelect = React.forwardRef<any, IColorSelectProps>(
-  (
-    {
-      label = "",
-      options = [],
-      defaultValue = "",
-      errorText = "",
-      onChange = () => {},
-      ...props
-    },
-    ref
-  ) => {
-    const [selectFocus, setSelectFocus] = useState(true);
-    const [color, setColor] = useState("transparent");
+const ColorSelect = ({
+  label = "",
+  options = [],
+  errorText = "",
+  value,
+  onChange = () => {},
+  ...props
+}: IColorSelectProps) => {
+  const [color, setColor] = useState<IColorSelectOptions>(
+    value || { label: "", value: "" }
+  );
 
-    const handleChange = (e: any) => {
-      let value = e.target.value;
-      setColor(value);
-      onChange(e);
-    };
-
-    useEffect(() => {
-      if (defaultValue) {
-        setColor(defaultValue);
-      }
-    }, [defaultValue]);
-
-    return (
-      <div className="w-full h-full flex items-center relative">
-        <div className="flex flex-1 flex-col mb-2 relative">
-          <label
-            htmlFor={props.id}
-            className="text-lg absolute text-gray-400 left-3 top--1/2 duration-300 m-0"
-            style={{
-              transform: selectFocus ? "translate(0, 0%)" : "translate(0, 40%)",
-              fontSize: selectFocus ? "0.95rem" : "1.125rem",
-            }}
-          >
-            {label}
-          </label>
-          <select
-            onChange={handleChange}
-            className="flex-grow border border-gray-400 rounded-lg text-md px-1.5 pb-1 pt-5 outline-1 outline-blue-100"
-            {...props}
-            onFocus={() => setSelectFocus(true)}
-            onBlur={(e) => e.target.value || setSelectFocus(false)}
-            ref={ref}
-          >
-            {options.map((option) => {
-              return (
-                <option value={option.value} key={option.value}>
-                  {option.label}
-                </option>
-              );
-            })}
-          </select>
-          {errorText && <ErrorMessage>{errorText}</ErrorMessage>}
-        </div>
+  return (
+    <div className="w-full h-full flex relative">
+      <SelectWithLabel
+        {...{ label, options, errorText, value }}
+        {...props}
+        value={color}
+        onChange={(e) => {
+          setColor(e.target.value);
+          onChange && onChange(e);
+        }}
+      />
+      <div className="w-12 h-12 ml-2 rounded-lg overflow-hidden mb-auto bg-grey-700 shadow-md">
         <div
-          className="w-11 h-11 ml-2 rounded-lg mb-2"
+          className="w-full h-full flex justify-center items-center"
           style={{
-            backgroundColor: color,
+            backgroundColor: color?.value || "",
           }}
-        ></div>
+        >
+          {!color?.value && (
+            <MdOutlineInvertColorsOff
+              fontSize={"25px"}
+              className={"text-white"}
+            />
+          )}
+        </div>
       </div>
-    );
-  }
-);
+    </div>
+  );
+};
+
+interface IColorSelectWithControlProps extends IColorSelectProps {
+  control?: any;
+  id?: string;
+  rules?: any;
+}
+
+export const ColorSelectWithControl = ({
+  control,
+  rules,
+  ...props
+}: IColorSelectWithControlProps) => {
+  return (
+    <Controller
+      control={control}
+      name={props?.id as string}
+      rules={rules}
+      render={({ field }) => <ColorSelect {...props} {...field} />}
+    />
+  );
+};
 
 export default ColorSelect;
