@@ -2,10 +2,10 @@ import React, { useEffect, useState } from "react";
 import { createTodo } from "../../redux/actions/todo.actions";
 import { useDispatch, useSelector } from "react-redux";
 import { PrimaryButton } from "../../styles/styles";
-import { useForm } from "react-hook-form";
+import { Controller, useForm } from "react-hook-form";
 import FormInput from "../FormInput/FormInput";
 import { ITodoState } from "@types";
-import Select from "../Select/Select";
+import { SelectWithLabelWithControl } from "../SelectWithLabel/SelectWithLabel";
 import { RootState } from "src/redux/store/store";
 import _ from "lodash";
 
@@ -32,13 +32,14 @@ const AddTodo = ({ setOpen }: IAddTodoProps) => {
     register,
     handleSubmit,
     formState: { errors },
+    control,
   } = useForm();
 
   //submit form
   const formSubmitHandler = ({ title, category }: Partial<ITodoState>) => {
     const data = {
       title,
-      category,
+      category: category.value,
     };
     console.log(data, "data");
     dispatch(createTodo(data));
@@ -68,26 +69,18 @@ const AddTodo = ({ setOpen }: IAddTodoProps) => {
         />
       </fieldset>
       <fieldset>
-        {_.isEmpty(todoCategoryList) ? (
-          <Select
-            disabled={true}
-            options={[{ label: "No categories", value: "" }]}
-            label="Category"
-            errorText={errors?.category?.message}
-            {...register("category", {
-              required: "Category is required",
-            })}
-          />
-        ) : (
-          <Select
-            options={[{ label: "", value: "" }, ...todoCategoryList]}
-            label="Category"
-            errorText={errors?.category?.message}
-            {...register("category", {
-              required: "Category is required",
-            })}
-          />
-        )}
+        <SelectWithLabelWithControl
+          options={
+            todoCategoryList && todoCategoryList[0]
+              ? [...todoCategoryList]
+              : [{ value: "", label: "" }]
+          }
+          control={control}
+          rules={{ required: "Category is required" }}
+          id="category"
+          label="Category"
+          errorText={errors?.category?.message}
+        />
       </fieldset>
       <PrimaryButton
         disabled={errors?.title || errors?.category}
